+++ src/main/java/org/owntracks/android/activities/ActivityExport.java
@@ -8,6 +8,7 @@ import android.content.Intent;
 import android.graphics.Color;
 import android.net.Uri;
 import android.os.Bundle;
+import android.support.design.widget.Snackbar;
 import android.support.v4.content.FileProvider;
 import android.support.v7.app.ActionBarActivity;
 import android.support.v7.app.AppCompatActivity;
@@ -41,18 +42,6 @@ public class ActivityExport extends ActivityBase {
     private static final String TAG = "ActivityExport";
 
     private static final String TEMP_FILE_NAME = "config.otrc";
-    private Switch includeConnection;
-    private Switch includeCredentials;
-    private Switch includeDeviceIdentification;
-    private Switch includeWaypoints;
-
-    boolean includeConnectionVal=true;
-    boolean includeCredentialsVal=false;
-    boolean includeDeviceIdentificationVal=false;
-    boolean includeWaypointsVal=true;
-    private TextView exportUsernameLabel;
-    private TextView exportIdentificationLabel;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -63,43 +52,6 @@ public class ActivityExport extends ActivityBase {
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setTitle(getResources().getString(R.string.export));
 
-        includeConnection = (Switch) findViewById(R.id.includeConnection);
-        includeCredentials = (Switch) findViewById(R.id.includeUsernamePassword);
-        includeDeviceIdentification = (Switch) findViewById(R.id.includeDeviceIdentification);
-        includeWaypoints = (Switch) findViewById(R.id.includeWaypoints);
-
-        exportUsernameLabel = (TextView) findViewById(R.id.exportUsernameLabel);
-        exportIdentificationLabel = (TextView) findViewById(R.id.exportIdentificationLabel);
-
-        setUsernameDeviceExport(true);
-
-        includeConnection.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                includeConnectionVal = isChecked;
-                setUsernameDeviceExport(includeConnectionVal);
-            }
-        });
-
-        includeCredentials.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                includeCredentialsVal = isChecked;
-            }
-        });
-
-        includeDeviceIdentification.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                includeDeviceIdentificationVal = isChecked;
-            }
-        });
-        includeWaypoints.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                includeWaypointsVal = isChecked;
-            }
-        });
     }
 
     @Override
@@ -125,50 +77,11 @@ public class ActivityExport extends ActivityBase {
 
 
 
-    private void setUsernameDeviceExport(boolean isChecked) {
-        includeCredentials.setEnabled(isChecked);
-        includeDeviceIdentification.setEnabled(isChecked);
-        int primaryColor = getResources().getColor(R.color.textPrimary);
-        int textColor = Color.argb(isChecked? 255 : 128,
-                Color.red(primaryColor),
-                Color.green(primaryColor),
-                Color.blue(primaryColor));
-
-        exportUsernameLabel.setTextColor(textColor);
-        exportIdentificationLabel.setTextColor(textColor);
-
-        // Uncheck Credentials and DeviceIdentification if Connection is not checked
-        if(!isChecked) {
-            includeCredentials.setChecked(false);
-            includeCredentialsVal =false;
-            includeDeviceIdentification.setChecked(false);
-            includeDeviceIdentificationVal = false;
-        }
-    }
-
     private void export() {
-        Log.v("Export", "Export includes: connection=" + includeConnection.isChecked() + ", username/password=" + includeCredentials.isChecked() + ", device identification=" + includeDeviceIdentification.isChecked() + ", waypoints=" + includeWaypoints.isChecked());
 
 
-
-        EnumSet<ConfigurationMessage.Includes> includes = EnumSet.noneOf(ConfigurationMessage.Includes.class);
-
-        if (includeConnection.isChecked() && includeConnection.isEnabled())
-            includes.add(ConfigurationMessage.Includes.CONNECTION);
-        if (includeCredentials.isChecked()&& includeCredentials.isEnabled())
-            includes.add(ConfigurationMessage.Includes.CREDENTIALS);
-        if (includeDeviceIdentification.isChecked() && includeCredentials.isEnabled())
-            includes.add(ConfigurationMessage.Includes.IDENTIFICATION);
-        if (includeWaypoints.isChecked())
-            includes.add(ConfigurationMessage.Includes.WAYPOINTS);
-
-        Log.v(TAG, "Export includes: " + includes);
-
-        ConfigurationMessage config = new ConfigurationMessage(includes);
+        ConfigurationMessage config = new ConfigurationMessage();
         Log.v("Export", "Config: \n" + config.toString());
-
-
-
 
         File cDir = getBaseContext().getCacheDir();
         File tempFile = new File(cDir.getPath() + "/" + TEMP_FILE_NAME) ;
@@ -185,17 +98,13 @@ public class ActivityExport extends ActivityBase {
             e.printStackTrace();
         }
         Uri configUri = FileProvider.getUriForFile(this, "org.owntracks.android.fileprovider", tempFile);
-        //configUri = Uri.parse(configUri.toString() + ".otrc");
-
-
 
         Intent sendIntent = new Intent();
         sendIntent.setAction(Intent.ACTION_SEND);
         //sendIntent.putExtra(Intent.EXTRA_TEXT, config.toString());
         sendIntent.putExtra(Intent.EXTRA_STREAM, configUri);
         sendIntent.setType("text/plain");
-
-
         startActivity(Intent.createChooser(sendIntent, getString(R.string.exportConfiguration)));
     }
+
 }